MixedModelsPermutations.jl API

Nonparametric Bootstrap

MixedModelsPermutations.nonparametricbootstrap — Function

nonparametricbootstrap([rng::AbstractRNG,] nsamp::Integer, m::LinearMixedModel;
                       use_threads=false)

Perform nsamp nonparametric bootstrap replication fits of m, returning a MixedModelBootstrap.

The default random number generator is Random.GLOBAL_RNG.

GeneralizedLinearMixedModel is currently unsupported.

Named Arguments

use_threads determines whether or not to use thread-based parallelism.

Note that use_threads=true may not offer a performance boost and may even decrease peformance if multithreaded linear algebra (BLAS) routines are available. In this case, threads at the level of the linear algebra may already occupy all processors/processor cores. There are plans to provide better support in coordinating Julia- and BLAS-level threads in the future.

Method

The method implemented here is based on the approach given in Section 3.2 of: Carpenter, J.R., Goldstein, H. and Rasbash, J. (2003). A novel bootstrap procedure for assessing the relationship between class size and achievement. Journal of the Royal Statistical Society: Series C (Applied Statistics), 52: 431-443. https://doi.org/10.1111/1467-9876.00415

Note that this method is not exported to match permute!.

Missing docstring.

Missing docstring for MixedModels.resample!. Check Documenter's build log for details.

Permutation Testing

MixedModelsPermutations.permutation — Function

permutation([rng::AbstractRNG,] nsamp::Integer, m::LinearMixedModel;
                       use_threads=false)

Perform nsamp nonparametric bootstrap replication fits of m, returning a MixedModelBootstrap.

The default random number generator is Random.GLOBAL_RNG.

GeneralizedLinearMixedModel is currently unsupported.

Named Arguments

use_threads determines whether or not to use thread-based parallelism.

Note that use_threads=true may not offer a performance boost and may even decrease peformance if multithreaded linear algebra (BLAS) routines are available. In this case, threads at the level of the linear algebra may already occupy all processors/processor cores. There are plans to provide better support in coordinating Julia- and BLAS-level threads in the future.

Method

The method implemented here is based on the approach given in

Note that this method is not exported to avoid a name collision with Base.permute!.

MixedModelsPermutations.permute! — Function

permute!([rng::AbstractRNG,] mod::LinearMixedModel,
          blups=ranef(mod), reterms=mod.reterms)

Simulate and install a new response via sign-flipped permutation of the residuals at the observational and sign-flipping of the conditional modes at group level.

Sign-flipped permutation of the residuals is similar to permuting the (fixed-effects) design matrix. Sign-flipping the conditional modes (random effects) preserves the correlation structure of the random effects, while also being equivalent to permutation via swapped labels for categorical variables.

Note

This method has serious limitations for singular models because permuting zeros

See also permutation, nonparametricbootstrap and resample!.

Reference

ter Braak?